Add /dumpstructure for easier creation of multiblock templates
The command takes 2 block positions, the padding mode and a boolean.
The first position is the left corner, the second is the right corner.
Padding mode determines where will empty space be added to accommodate
non-square structures (the multiblock template wants squares).
If the last parameter is "true" adjacent empty spots will be put on
one line.