Migrate reflection to ObfuscationReflectionHelper
Greatly simplifies reflection code and keeps it mappings-independent.
Also you no longer need to copypaste the same snippet whenever you need
to reflect into something.

The one advantage of the previous approach was caching fields, but we
didn't do that at all and the code doesn't run very often anyways...
and if we decide to do that, we can just use ReflectionHelper instead.