Fix ore ingr. conversion only grabbing the first matching entry, build
While this is not the perfect solution, finding out which ore dict entry
was used to create the ingredient would require coremodding as this data
is not stored. This is not an issue in 1.14 as the tag system leaves the
used tags inside the ingredient, which can be accessed with reflection,
but this will have to do here.